# docs/30-COMPONENTES/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agnostic Orchestration Platform Control Plane",
    "version": "0.1.0"
  },
  "paths": {
...
    "/squads/{squad_id}/topology": {
      "post": {
        "tags": [
          "squads"
        ],
        "summary": "Save Topology",
        "operationId": "save_topology_squads__squad_id__topology_post",
        "parameters": [
          {
            "name": "squad_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Squad Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopologySaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Save Topology Squads  Squad Id  Topology Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "squads"
        ],
        "summary": "Get Topology",
        "operationId": "get_topology_squads__squad_id__topology_get",
        "parameters": [
          {
            "name": "squad_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Squad Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Topology Squads  Squad Id  Topology Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TopologyEdgeRequest": {
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "title": "Source"
          },
          "target": {
            "type": "string",
            "minLength": 1,
            "title": "Target"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "source",
          "target"
        ],
        "title": "TopologyEdgeRequest"
      },
      "TopologyNodeRequest": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id"
        ],
        "title": "TopologyNodeRequest"
      },
      "TopologySaveRequest": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/TopologyNodeRequest"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/TopologyEdgeRequest"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nodes",
          "edges"
        ],
        "title": "TopologySaveRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}